Removed old php + updated html, css
-removed php files that arent used anymore
-updated html class names for the legend
