+++ static/css/legendStyle.css
@@ -53,6 +53,8 @@
     margin-left: 0;
     text-align: left;
 }
+
+/*styles the seperating bars*/
 .legend hr{
     padding:0%;
     margin-bottom: 0%;
@@ -168,7 +170,8 @@
   -webkit-transform: rotate(180deg);
 }
 
-.drop-down-arrow-MD{
+/*creates arrow that toggles Use Type visibility*/
+.drop-down-arrow-UseType{
   width: 0;
   height: 0;
   border-left: 7px solid transparent;
@@ -180,23 +183,31 @@
   margin-right: 3%;
   margin-bottom: 1%;
 }
-  /*rotates drop-down arrow 90 degrees*/
-  .rotate-drop-down-180-MD {
+  
+/*rotates UseType drop-down arrow */
+.rotate-arrow-UseType {
   transform: rotate(-90deg); 
   -webkit-transform: rotate(-90deg);
 }
 
 /* toggles display visibility for MD data when clicking dropdown arrow */
-.drop-down-MD.hidden{
+.drop-down-UseTypes.hidden{
   display:none;
 }
 
+/* toggles display visibility for VA data when clicking dropdown arrow */
+.drop-down-States.hidden{
+  display:none;
+}
+
+/*positioning for legend headers(States & Use Types)*/
 .drop-down-header > *{
   width:100%;
   margin-top: 15px;
 }
 
-.drop-down-arrow-VA{
+/*creates arrow that toggles States visibility*/
+.drop-down-arrow-States{
   width: 0;
   height: 0;
   border-left: 7px solid transparent;
@@ -209,24 +220,22 @@
   margin-bottom: 1%;
 }
 
-.drop-down-arrow-VA:hover{
+/*on hover turns States dropdown arrow black*/
+.drop-down-arrow-States:hover{
   border-top: 7px solid black;
 }
 
-.drop-down-arrow-MD:hover{
+/*on hover turns UseType dropdown arrow black*/
+.drop-down-arrow-UseType:hover{
   border-top: 7px solid black;
 }
 
-  /*rotates drop-down arrow 90 degrees */
-  .rotate-drop-down-180-VA {
+/*rotates States drop-down arrow 90 degrees */
+.rotate-arrow-States {
   transform: rotate(-90deg); 
   -webkit-transform: rotate(-90deg);
 }
 
-/* toggles display visibility for VA data when clicking dropdown arrow */
-.drop-down-VA.hidden{
-  display:none;
-}
