+++ static/css/popupStyle.css
@@ -31,7 +31,7 @@
     border: 1px solid #ccc;
 }
 
-  /* sets title color to black*/
+/* sets title color to black*/
 .info-window strong{
     color: #333;
 }
@@ -39,8 +39,8 @@
 /*styling for popup close button 'X' */
 #close-button {
     position: absolute;
-    top: 10px;
-    right: 10px;
+    top: 15px;
+    right: 15px;
     cursor: pointer;
     color: white;
 }
@@ -58,6 +58,7 @@
     cursor: pointer;
 }
 
+/*styling for view more button*/
 #view-more-button{
     padding: 5px 15px; 
     border-radius: 20px;
@@ -92,9 +93,7 @@
     color: rgb(70, 86, 126);
     margin: 20px;
 }
-.gm-style-iw button:focus {
-    outline: 0;
-}
+
 .header-paragraph p{
     position: relative;
     text-align: left;
@@ -102,7 +101,7 @@
     color: black;
     margin-top: 0;
 }
- 
+
 .text{
     font-size: 16px;
     font-family:system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
@@ -112,22 +111,3 @@
     border: none;
     background: transparent;
 }
-
-
-/* remove info-window X button
-button.gm-ui-hover-effect {
-    visibility: hidden;
-}
-
-
-/* Changes info-window background color
-.gm-style .gm-style-iw-d::-webkit-scrollbar-track, 
-.gm-style .gm-style-iw-d::-webkit-scrollbar-track-piece,
-.gm-style .gm-style-iw-c,
-.gm-style .gm-style-iw-t::after { 
-  background: rgb(70, 86, 126);
-}
-
-.gm-style .gm-style-iw-tc::after {   background: rgb(70, 86, 126); }
-
-*/
