+++ static/css/searchbar.css
@@ -7,7 +7,7 @@
   padding: 10px;
   border-radius: 23px;
   background: #ffffff;
-  margin-left: 15%;
+  margin-left: 2%;
   margin-top: .5%;
   position: absolute;
   z-index: 1;
